feat: implement tablecodegen package
Decoupled transaction types to enable reading across tables within a
transaction, with a convenience table type wrapping spanner.Client for
the common use cases.

The API and behavior is intended to be as close to Spanner as possible,
i.e. no error wrapping or error modification.

Interleaved table support is under development.