feat: simplify query API surface
Single entrypoint to queries: `Query()`.

Rows and interleaved rows have been merged. Mutation functions
simplified. (It's now up to the caller to call the `Mutate` method for
each interleaved row).

An upcoming PR will add support for arbitrary depth interleaving, and an
update to the Interleaved queries making it possible to specify which
interleaved tables to include in the results.