feat!: transpile filters to expressions with parameters
Instead of transpiling filters to a single spansql.BoolExpr with
spansql.[...]Literal, filters are transpiled to a spansql.BoolExpr with
every literal returned as a parameter.

This fixes potential security issue where malicous filters are
transpiled directly to SQL without proper escaping.

Parameters are automatically named using a counter based on order of
appearance in the expr.Expr AST. This could be prettier, but atleast
provides a deterministic naming which *could* improve ability to cache
the filter (this is an assumption).