feat: initial implementation of spaniam.Server
Implements Spanner storage for standard IAM policies.

Role definitions in this initial implementation are assumed to be static
and not managed by Spanner. This is to keep the implementation as simple
as possible, especially in the case where built-in roles and permissions
are declared as part of the gRPC service API.