feat: don't consider soft delete for interleaved rows
This is too hard to get right for all use cases. The simplest thing to
reason about is to only consider soft deletes for the top-level table
(because that's the one that determines pagination) and let the caller
filter out interleaved results as required.

BREAKING CHANGE: Soft-deleted interleaved rows will now always be
                 returned.