feat: do not run interleaved queries with empty keys
If the user attempts to call a generated List method without supplying a
key (with the purpose of returning the entire table), we get an empty
slice as a reply if no data exists in the database.
If the same is done on a table which has a interleaved child and the
user asks for interleaved data by setting the RowsQuery field to true,
we attempt to perform a query on the interleaved table without having a
key to perform such a query (since there is no data returned from the
parent List) and we get back an `Unimplemented` error from the spanner
library with the error message `Cloud Spanner does not support reading
no keys`.

This fix makes it so that if the list query returned no keys, we do not
query the interleaved child table and just return an empty slice