refactor(databasegen): make interleaved table reads independent
Preparation for running all reads for interleaved tables in parallell.

Currently the order in which child tables are read matters, which
prevents the queries from running in parallell.

This commit appends child table rows to a slice during read, and then
performs the rollup as a final step, instead of doing rollup as part of
the read.